apisample14.html

<!DOCTYPE HTML>
<title>Example with a dedicated worker</title>
<script src="testharness.js"></script>
<script src="testharnessreport.js"></script>
<script>
test(function(t) {
  assert_true('Worker' in self, "Browser should support Workers");
}, "Browser supports Workers");

fetch_tests_from_worker(new Worker('apisample-worker.js'));

test(function(t) {
  assert_false(false, "False is false");
}, "Test running on main document. " +
"There should be an additional test named \'Test running on worker\'.");
</script>
